Add support for legacy android camera api
