+++ FluStudy_us/android/app/src/main/java/host/exp/exponent/CameraController.java
@@ -1,53 +1,27 @@
-// Copyright (c) 2019 by Audere
-//
-// Use of this source code is governed by an MIT-style license that
-// can be found in the LICENSE file distributed with this file.
-
 package host.exp.exponent;
 
 import android.app.Activity;
 import android.content.Context;
-import android.content.res.Configuration;
-import android.graphics.ImageFormat;
-import android.graphics.Matrix;
-import android.graphics.RectF;
-import android.graphics.SurfaceTexture;
 import android.hardware.camera2.CameraAccessException;
-import android.hardware.camera2.CameraCaptureSession;
 import android.hardware.camera2.CameraCharacteristics;
-import android.hardware.camera2.CameraDevice;
-import android.hardware.camera2.CameraMetadata;
-import android.hardware.camera2.CaptureRequest;
-import android.hardware.camera2.CaptureResult;
-import android.hardware.camera2.TotalCaptureResult;
-import android.hardware.camera2.params.MeteringRectangle;
+import android.hardware.camera2.CameraManager;
 import android.hardware.camera2.params.StreamConfigurationMap;
-import android.media.ImageReader;
-import android.os.Handler;
-import android.os.HandlerThread;
 import android.text.TextUtils;
 import android.util.Log;
 import android.util.Size;
-import android.util.SparseIntArray;
-import android.view.Surface;
-import android.view.TextureView;
-import android.widget.Toast;
 
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.Comparator;
 import java.util.List;
-import java.util.concurrent.Semaphore;
-import java.util.concurrent.TimeUnit;
 
 import host.exp.exponent.customview.AutoFitTextureView;
 
-public class CameraController {
+public abstract class CameraController {
     private static final String TAG = "CameraController";
-    private Activity activity;
-    private boolean flashEnabled = false;
-    private boolean isMeteringAreaAFSupported;
+    protected final Activity activity;
+    protected boolean flashEnabled;
 
     /**
      * The camera preview size will be chosen to be the smallest frame by pixel size capable of
@@ -55,232 +29,124 @@ public class CameraController {
      */
     private static final int MINIMUM_PREVIEW_SIZE = 320;
 
-    /** Conversion from screen rotation to JPEG orientation. */
-    private static final SparseIntArray ORIENTATIONS = new SparseIntArray();
-
-    /**
-     * Camera state: Showing camera preview.
-     */
-    private static final int STATE_PREVIEW = 0;
-
-    /**
-     * Camera state: Waiting for the focus to be locked.
-     */
-    private static final int STATE_WAITING_LOCK = 1;
-
-    /**
-     * Camera state: Waiting for the exposure to be precapture state.
-     */
-    private static final int STATE_WAITING_PRECAPTURE = 2;
-
-    /**
-     * Camera state: Waiting for the exposure state to be something other than precapture.
-     */
-    private static final int STATE_WAITING_NON_PRECAPTURE = 3;
-
-    /**
-     * Camera state: Picture was taken.
-     */
-    private static final int STATE_PICTURE_TAKEN = 4;
-
-    /**
-     * Camera state: waiting for the focus state to be reset.
-     */
-    private static final int STATE_WAITING_AF_RESET = 5;
-
-    static {
-        ORIENTATIONS.append(Surface.ROTATION_0, 90);
-        ORIENTATIONS.append(Surface.ROTATION_90, 0);
-        ORIENTATIONS.append(Surface.ROTATION_180, 270);
-        ORIENTATIONS.append(Surface.ROTATION_270, 180);
-    }
-
-    /** A {@link Semaphore} to prevent the app from exiting before closing the camera. */
-    private final Semaphore cameraOpenCloseLock = new Semaphore(1);
-    /** A {@link ImageReader.OnImageAvailableListener} to receive frames as they are available. */
-    private final ImageReader.OnImageAvailableListener imageListener;
-    /** A {@link ImageReader.OnImageAvailableListener} to receive the still image. */
-    private final ImageReader.OnImageAvailableListener stillImageListener;
-    /** The input size in pixels desired by TensorFlow (width and height of a square bitmap). */
-    private final Size inputSize;
-
-    private final ConnectionCallback cameraConnectionCallback;
-    /** ID of the current {@link CameraDevice}. */
-    private String cameraId;
+    /** ID of the current camera device. */
+    protected final String cameraId;
     /** An {@link AutoFitTextureView} for camera preview. */
-    private AutoFitTextureView textureView;
-    /** A {@link CameraCaptureSession } for camera preview. */
-    private CameraCaptureSession captureSession;
-    /** A reference to the opened {@link CameraDevice}. */
-    private CameraDevice cameraDevice;
-    /** The rotation in degrees of the camera sensor from the display. */
-    private Integer sensorOrientation;
-    /** The {@link Size} of camera preview. */
-    private Size previewSize;
-    /** The {@link Size} of the high quality still. */
-    private Size stillSize;
-    /** An additional thread for running tasks that shouldn't block the UI. */
-    private HandlerThread backgroundThread;
-    /** A {@link Handler} for running tasks in the background. */
-    private Handler backgroundHandler;
+    protected final AutoFitTextureView textureView;
+    protected final ConnectionCallback cameraConnectionCallback;
+    protected final Size desiredSize;
 
-    /** An {@link ImageReader} that handles preview frame capture. */
-    private ImageReader previewReader;
-    /** An {@link ImageReader} that handles stills. */
-    private ImageReader stillReader;
-    /** {@link CaptureRequest.Builder} for the camera preview */
-    private CaptureRequest.Builder previewRequestBuilder;
-    /** {@link CaptureRequest} generated by {@link #previewRequestBuilder} */
-    private CaptureRequest previewRequest;
+    /** An ImageListener that handles preview frame capture. */
+    protected final DetectorView.PreviewImageListener imageListener;
+    /** An ImageListener that handles stills. */
+    protected final DetectorView.StillImageListener stillImageListener;
 
-    private int state = STATE_PREVIEW;
-    private int focusResets = 0;
-    private final int MAX_FOCUS_RESETS = 5;
-
-    /**
-     * A {@link CameraCaptureSession.CaptureCallback} that handles events related to JPEG capture.
-     */
-    private CameraCaptureSession.CaptureCallback captureCallback
-            = new CameraCaptureSession.CaptureCallback() {
-
-        private void process(CaptureResult result) {
-            switch (state) {
-                case STATE_PREVIEW: {
-                    // We have nothing to do when the camera preview is working normally.
-                    break;
-                }
-                case STATE_WAITING_AF_RESET: {
-                    lockFocus();
-                    break;
-                }
-                case STATE_WAITING_LOCK: {
-                    Integer afState = result.get(CaptureResult.CONTROL_AF_STATE);
-                    if (afState == null) {
-                        captureStillPicture();
-                    } else if (CaptureResult.CONTROL_AF_STATE_NOT_FOCUSED_LOCKED == afState &&
-                            focusResets < MAX_FOCUS_RESETS) {
-                        focusResets++;
-                        resetFocus();
-                    } else if ((CaptureResult.CONTROL_AF_STATE_NOT_FOCUSED_LOCKED == afState ||
-                            CaptureResult.CONTROL_AF_STATE_FOCUSED_LOCKED == afState)) {
-                        focusResets = 0;
-                        // CONTROL_AE_STATE can be null on some devices
-                        Integer aeState = result.get(CaptureResult.CONTROL_AE_STATE);
-                        if (aeState == null ||
-                                aeState == CaptureResult.CONTROL_AE_STATE_CONVERGED) {
-                            state = STATE_PICTURE_TAKEN;
-                            captureStillPicture();
-                        } else {
-                            runPrecaptureSequence();
-                        }
-                    }
-                    break;
-                }
-                case STATE_WAITING_PRECAPTURE: {
-                    // CONTROL_AE_STATE can be null on some devices
-                    Integer aeState = result.get(CaptureResult.CONTROL_AE_STATE);
-                    if (aeState == null ||
-                            aeState == CaptureResult.CONTROL_AE_STATE_PRECAPTURE ||
-                            aeState == CaptureRequest.CONTROL_AE_STATE_FLASH_REQUIRED) {
-                        state = STATE_WAITING_NON_PRECAPTURE;
-                    }
-                    break;
-                }
-                case STATE_WAITING_NON_PRECAPTURE: {
-                    // CONTROL_AE_STATE can be null on some devices
-                    Integer aeState = result.get(CaptureResult.CONTROL_AE_STATE);
-                    if (aeState == null || aeState != CaptureResult.CONTROL_AE_STATE_PRECAPTURE) {
-                        state = STATE_PICTURE_TAKEN;
-                        captureStillPicture();
-                    }
-                    break;
-                }
-            }
-        }
-
-        @Override
-        public void onCaptureProgressed(CameraCaptureSession session,
-                                        CaptureRequest request,
-                                        CaptureResult partialResult) {
-            process(partialResult);
-        }
-
-        @Override
-        public void onCaptureCompleted(CameraCaptureSession session,
-                                       CaptureRequest request,
-                                       TotalCaptureResult result) {
-            process(result);
-        }
-
-    };
-
-
-    /** {@link CameraDevice.StateCallback} is called when {@link CameraDevice} changes its state. */
-    private final CameraDevice.StateCallback stateCallback =
-            new CameraDevice.StateCallback() {
-                @Override
-                public void onOpened(final CameraDevice cd) {
-                    // This method is called when the camera is opened.  We start camera preview here.
-                    cameraOpenCloseLock.release();
-                    cameraDevice = cd;
-                    createCameraPreviewSession();
-                }
-
-                @Override
-                public void onDisconnected(final CameraDevice cd) {
-                    cameraOpenCloseLock.release();
-                    cd.close();
-                    cameraDevice = null;
-                }
-
-                @Override
-                public void onError(final CameraDevice cd, final int error) {
-                    cameraOpenCloseLock.release();
-                    cd.close();
-                    cameraDevice = null;
-                }
-            };
-    /**
-     * {@link TextureView.SurfaceTextureListener} handles several lifecycle events on a {@link
-     * TextureView}.
-     */
-    private final TextureView.SurfaceTextureListener surfaceTextureListener =
-            new TextureView.SurfaceTextureListener() {
-                @Override
-                public void onSurfaceTextureAvailable(
-                        final SurfaceTexture texture, final int width, final int height) {
-                    openCamera(width, height);
-                }
-
-                @Override
-                public void onSurfaceTextureSizeChanged(
-                        final SurfaceTexture texture, final int width, final int height) {
-                    configureTransform(width, height);
-                }
-
-                @Override
-                public boolean onSurfaceTextureDestroyed(final SurfaceTexture texture) {
-                    closeCamera();
-                    return true;
-                }
-
-                @Override
-                public void onSurfaceTextureUpdated(final SurfaceTexture texture) {}
-            };
-
-    public CameraController(
-            final Activity activity,
-            final AutoFitTextureView textureView,
-            final ConnectionCallback callback,
-            final ImageReader.OnImageAvailableListener imageListener,
-            final ImageReader.OnImageAvailableListener stillImageListener,
-            final Size inputSize) {
+    protected CameraController(final Activity activity,
+                                 final AutoFitTextureView textureView,
+                                 final ConnectionCallback callback,
+                                 final DetectorView.PreviewImageListener imageListener,
+                                 final DetectorView.StillImageListener stillImageListener,
+                                 final Size desiredSize,
+                                 final String cameraId) {
         this.activity = activity;
         this.textureView = textureView;
         this.cameraConnectionCallback = callback;
         this.imageListener = imageListener;
         this.stillImageListener = stillImageListener;
-        this.inputSize = inputSize;
+        this.desiredSize = desiredSize;
+        this.cameraId = cameraId;
+    }
+
+    public abstract void onResume();
+
+    public abstract void onPause();
+
+    public abstract void captureStill();
+
+    public abstract void setFlashEnabled(boolean flashEnabled);
+
+    public static CameraController getCamera(final Activity activity,
+                                               final AutoFitTextureView textureView,
+                                               final ConnectionCallback callback,
+                                               final DetectorView.PreviewImageListener imageListener,
+                                               final DetectorView.StillImageListener stillImageListener,
+                                               final Size desiredSize) {
+
+        boolean useCamera2API = false;
+        String selectedCameraID = null;
+
+        final CameraManager manager = (CameraManager) activity.getSystemService(Context.CAMERA_SERVICE);
+        try {
+            for (final String cameraId : manager.getCameraIdList()) {
+                final CameraCharacteristics characteristics = manager.getCameraCharacteristics(cameraId);
+
+                // We don't use a front facing camera in this sample.
+                final Integer facing = characteristics.get(CameraCharacteristics.LENS_FACING);
+                if (facing != null && facing == CameraCharacteristics.LENS_FACING_FRONT) {
+                    continue;
+                }
+
+                final StreamConfigurationMap map =
+                        characteristics.get(CameraCharacteristics.SCALER_STREAM_CONFIGURATION_MAP);
+
+                if (map == null) {
+                    continue;
+                }
+
+                useCamera2API =
+                        (facing == CameraCharacteristics.LENS_FACING_EXTERNAL)
+                                || isHardwareLevelSupported(
+                                characteristics, CameraCharacteristics.INFO_SUPPORTED_HARDWARE_LEVEL_FULL);
+
+                selectedCameraID = cameraId;
+                break;
+            }
+
+            if (useCamera2API) {
+                return new RDTCamera(activity, textureView, callback, imageListener,
+                        stillImageListener, desiredSize, selectedCameraID);
+            } else {
+                return new LegacyRDTCamera(activity, textureView, callback, imageListener,
+                        stillImageListener, desiredSize, selectedCameraID);
+            }
+        } catch (CameraAccessException e) {
+            Log.e(TAG, "Not allowed to access camera");
+        }
+        return null;
+    }
+
+    // Returns true if the device supports the required hardware level, or better.
+    private static boolean isHardwareLevelSupported(CameraCharacteristics c, int requiredLevel) {
+        final int[] sortedHwLevels = {
+                CameraCharacteristics.INFO_SUPPORTED_HARDWARE_LEVEL_LEGACY,
+                CameraCharacteristics.INFO_SUPPORTED_HARDWARE_LEVEL_EXTERNAL,
+                CameraCharacteristics.INFO_SUPPORTED_HARDWARE_LEVEL_LIMITED,
+                CameraCharacteristics.INFO_SUPPORTED_HARDWARE_LEVEL_FULL,
+                CameraCharacteristics.INFO_SUPPORTED_HARDWARE_LEVEL_3
+        };
+        int deviceLevel = c.get(CameraCharacteristics.INFO_SUPPORTED_HARDWARE_LEVEL);
+        Log.d(TAG, "Device hardware supported level: " + deviceLevel);
+        Log.d(TAG, "All levels: " + Arrays.toString(sortedHwLevels));
+        if (requiredLevel == deviceLevel) {
+            return true;
+        }
+
+        for (int sortedlevel : sortedHwLevels) {
+            if (sortedlevel == requiredLevel) {
+                return true;
+            } else if (sortedlevel == deviceLevel) {
+                return false;
+            }
+        }
+        return false; // Should never reach here
+    }
+
+    /**
+     * Callback for Activities to use to initialize their data once the selected preview size is
+     * known.
+     */
+    public interface ConnectionCallback {
+        void onPreviewSizeChosen(Size previewSize, Size stillSize, int cameraRotation, boolean supportsTorchMode);
     }
 
     /**
@@ -324,9 +190,9 @@ public class CameraController {
 
         if (bigEnough.size() == 0) {
             Log.e(
-                TAG,
-                "No preview size large enough, using "
-                    + choices[0].getWidth() + "x" + choices[0].getHeight()
+                    TAG,
+                    "No preview size large enough, using "
+                            + choices[0].getWidth() + "x" + choices[0].getHeight()
             );
             return choices[0];
         }
@@ -344,443 +210,21 @@ public class CameraController {
 
         // Second try to find one that matches reversed aspect ratio
         for (final Size option : bigEnough) {
-          if (option.getHeight() * height == option.getWidth() * width) {
-            Log.i(TAG, "Matched inverse aspect ratio: " + option.getWidth() + "x" + option.getHeight());
-            return option;
-          }
+            if (option.getHeight() * height == option.getWidth() * width) {
+                Log.i(TAG, "Matched inverse aspect ratio: " + option.getWidth() + "x" + option.getHeight());
+                return option;
+            }
         }
 
         Size chosen = bigEnough.get(0);
         Log.w(
-            TAG,
-            "Large enough, but could not match aspect ratio, using "
-                + chosen.getWidth() + "x" + chosen.getHeight()
+                TAG,
+                "Large enough, but could not match aspect ratio, using "
+                        + chosen.getWidth() + "x" + chosen.getHeight()
         );
         return chosen;
     }
 
-    /**
-     * Shows a {@link Toast} on the UI thread.
-     *
-     * @param text The message to show
-     */
-    private void showToast(final String text) {
-        if (activity != null) {
-            activity.runOnUiThread(
-                    new Runnable() {
-                        @Override
-                        public void run() {
-                            Toast.makeText(activity, text, Toast.LENGTH_SHORT).show();
-                        }
-                    });
-        }
-    }
-
-    public void onResume() {
-        Log.i(TAG, "resume");
-        startBackgroundThread();
-
-        // When the screen is turned off and turned back on, the SurfaceTexture is already
-        // available, and "onSurfaceTextureAvailable" will not be called. In that case, we can open
-        // a camera and start preview from here (otherwise, we wait until the surface is ready in
-        // the SurfaceTextureListener).
-        if (textureView.isAvailable()) {
-            openCamera(textureView.getWidth(), textureView.getHeight());
-        } else {
-            textureView.setSurfaceTextureListener(surfaceTextureListener);
-        }
-    }
-
-    public void onPause() {
-        Log.i(TAG, "pause()");
-        closeCamera();
-        stopBackgroundThread();
-    }
-
-    public void setCamera(String cameraId) {
-        this.cameraId = cameraId;
-    }
-
-    /** Sets up member variables related to camera. */
-    private void setUpCameraOutputs() {
-        boolean supportsTorchMode = false;
-        final android.hardware.camera2.CameraManager manager = (android.hardware.camera2.CameraManager) activity.getSystemService(Context.CAMERA_SERVICE);
-        try {
-            final CameraCharacteristics characteristics = manager.getCameraCharacteristics(cameraId);
-
-            supportsTorchMode = characteristics.get(CameraCharacteristics.FLASH_INFO_AVAILABLE);
-            isMeteringAreaAFSupported = characteristics.get(CameraCharacteristics.CONTROL_MAX_REGIONS_AF) >= 1;
-            final StreamConfigurationMap map =
-                    characteristics.get(CameraCharacteristics.SCALER_STREAM_CONFIGURATION_MAP);
-
-            sensorOrientation = characteristics.get(CameraCharacteristics.SENSOR_ORIENTATION);
-            // Danger, W.R.! Attempting to use too large a preview size could  exceed the camera
-            // bus' bandwidth limitation, resulting in gorgeous previews but the storage of
-            // garbage capture data.
-            previewSize =
-                    chooseOptimalSize(
-                            map.getOutputSizes(SurfaceTexture.class),
-                            inputSize.getWidth(),
-                            inputSize.getHeight());
-            stillSize = Collections.max(Arrays.asList(map.getOutputSizes(ImageFormat.YUV_420_888)), new CompareSizesByArea());
-
-            // We fit the aspect ratio of TextureView to the size of preview we picked.
-            final int orientation = activity.getResources().getConfiguration().orientation;
-            if (orientation == Configuration.ORIENTATION_LANDSCAPE) {
-                textureView.setAspectRatio(previewSize.getWidth(), previewSize.getHeight());
-            } else {
-                textureView.setAspectRatio(previewSize.getHeight(), previewSize.getWidth());
-            }
-        } catch (final CameraAccessException e) {
-            Log.e(TAG, e.toString());
-        } catch (final NullPointerException e) {
-            // Currently an NPE is thrown when the Camera2API is used but not supported on the
-            // device this code runs.
-            showToast(activity.getString(R.string.camera_error));
-            throw e;
-            // throw new RuntimeException(activity.getString(R.string.camera_error));
-        }
-
-        cameraConnectionCallback.onPreviewSizeChosen(previewSize, stillSize, sensorOrientation, supportsTorchMode);
-    }
-
-    /** Opens the camera specified by cameraId. */
-    private void openCamera(final int width, final int height) {
-        setUpCameraOutputs();
-        configureTransform(width, height);
-        final android.hardware.camera2.CameraManager manager = (android.hardware.camera2.CameraManager) activity.getSystemService(Context.CAMERA_SERVICE);
-        try {
-            if (!cameraOpenCloseLock.tryAcquire(2500, TimeUnit.MILLISECONDS)) {
-                throw new RuntimeException("Time out waiting to lock camera opening.");
-            }
-            try {
-                manager.openCamera(cameraId, stateCallback, backgroundHandler);
-            } catch (SecurityException e) {
-                Log.e(TAG, "No permission to use camera");
-            }
-        } catch (final CameraAccessException e) {
-            Log.e(TAG, e.toString());
-        } catch (final InterruptedException e) {
-            throw new RuntimeException("Interrupted while trying to lock camera opening.", e);
-        }
-    }
-
-    /** Closes the current {@link CameraDevice}. */
-    private void closeCamera() {
-        Log.d(TAG, "close camera");
-        try {
-            cameraOpenCloseLock.acquire();
-            if (null != captureSession) {
-                captureSession.stopRepeating();
-                captureSession.close();
-                captureSession = null;
-            }
-            if (null != cameraDevice) {
-                cameraDevice.close();
-                cameraDevice = null;
-            }
-            if (null != previewReader) {
-                previewReader.close();
-                previewReader = null;
-            }
-
-            if (null != stillReader) {
-                stillReader.close();
-                stillReader = null;
-            }
-        } catch (CameraAccessException e) {
-            e.printStackTrace();
-        } catch (final InterruptedException e) {
-            throw new RuntimeException("Interrupted while trying to lock camera closing.", e);
-        } finally {
-            cameraOpenCloseLock.release();
-        }
-    }
-
-    /** Starts a background thread and its {@link Handler}. */
-    private void startBackgroundThread() {
-        backgroundThread = new HandlerThread("ImageListener");
-        backgroundThread.start();
-        backgroundHandler = new Handler(backgroundThread.getLooper());
-    }
-
-    /** Stops the background thread and its {@link Handler}. */
-    private void stopBackgroundThread() {
-        if (backgroundThread != null) {
-            backgroundThread.quitSafely();
-            backgroundThread = null;
-            backgroundHandler = null;
-        }
-    }
-
-    /** Creates a new {@link CameraCaptureSession} for camera preview. */
-    private void createCameraPreviewSession() {
-        try {
-            final SurfaceTexture texture = textureView.getSurfaceTexture();
-            assert texture != null;
-
-            // We configure the size of default buffer to be the size of camera preview we want.
-            texture.setDefaultBufferSize(previewSize.getWidth(), previewSize.getHeight());
-
-            // This is the output Surface we need to start preview.
-            final Surface surface = new Surface(texture);
-
-            // We set up a CaptureRequest.Builder with the output Surface.
-            previewRequestBuilder = cameraDevice.createCaptureRequest(CameraDevice.TEMPLATE_PREVIEW);
-
-            previewRequestBuilder.addTarget(surface);
-
-            Log.i(TAG, "Opening camera preview: " + previewSize.getWidth() + "x" + previewSize.getHeight());
-
-            // Create the reader for the preview frames.
-            previewReader =
-                    ImageReader.newInstance(
-                            previewSize.getWidth(), previewSize.getHeight(), ImageFormat.YUV_420_888, 2);
-            previewReader.setOnImageAvailableListener(imageListener, backgroundHandler);
-            previewRequestBuilder.addTarget(previewReader.getSurface());
-
-            stillReader = ImageReader.newInstance(stillSize.getWidth(), stillSize.getHeight(), ImageFormat.YUV_420_888, 2);
-            stillReader.setOnImageAvailableListener(stillImageListener, backgroundHandler);
-
-            // Here, we create a CameraCaptureSession for camera preview.
-            cameraDevice.createCaptureSession(
-                    Arrays.asList(surface, previewReader.getSurface(), stillReader.getSurface()),
-                    new CameraCaptureSession.StateCallback() {
-
-                        @Override
-                        public void onConfigured(final CameraCaptureSession cameraCaptureSession) {
-                            // The camera is already closed
-                            if (null == cameraDevice) {
-                                return;
-                            }
-
-                            // When the session is ready, we start displaying the preview.
-                            captureSession = cameraCaptureSession;
-
-                            try {
-                                // Auto focus should be continuous for camera preview.
-                                previewRequestBuilder.set(
-                                        CaptureRequest.CONTROL_AF_MODE,
-                                        CaptureRequest.CONTROL_AF_MODE_CONTINUOUS_PICTURE);
-
-
-                                setFlashOnRequestBuilder(previewRequestBuilder);
-
-                                previewRequestBuilder.setTag("PREVIEW");
-
-                                // Finally, we start displaying the camera preview.
-                                previewRequest = previewRequestBuilder.build();
-                                captureSession.setRepeatingRequest(
-                                        previewRequest, captureCallback, backgroundHandler);
-                            } catch (final CameraAccessException e) {
-                                Log.e(TAG, e.toString());
-                            }
-                        }
-
-                        @Override
-                        public void onConfigureFailed(final CameraCaptureSession cameraCaptureSession) {
-                            Log.d(TAG, "Failed to configure capture session");
-                        }
-                    },
-                    null);
-        } catch (final CameraAccessException e) {
-            Log.e(TAG, e.toString());
-        }
-    }
-
-    public void captureStill() {
-        if (cameraDevice == null) {
-            return;
-        }
-        resetFocus();
-    }
-
-    private void resetFocus() {
-        if (captureSession == null) {
-            return;
-        }
-        Log.d(TAG, "reset focus");
-        try {
-            // Stop existing repeating request
-            captureSession.stopRepeating();
-
-            // Turn off continuous capture mode
-            previewRequestBuilder.set(CaptureRequest.CONTROL_AF_TRIGGER, CameraMetadata.CONTROL_AF_TRIGGER_CANCEL);
-            state = STATE_WAITING_AF_RESET;
-            previewRequestBuilder.setTag("RESET_FOCUS");
-            captureSession.capture(previewRequestBuilder.build(), captureCallback, backgroundHandler);
-        } catch (CameraAccessException e) {
-            e.printStackTrace();
-        }
-    }
-
-    /**
-     * Lock the focus as the first step for a still image capture.
-     */
-    private void lockFocus() {
-        if (captureSession == null) {
-            return;
-        }
-        Log.d(TAG, "lockfocus");
-        try {
-            // Add a new AF trigger with focus region
-            previewRequestBuilder.set(CaptureRequest.CONTROL_AF_MODE,
-                    CaptureRequest.CONTROL_AF_MODE_CONTINUOUS_PICTURE);
-            previewRequestBuilder.set(CaptureRequest.CONTROL_AF_TRIGGER, CameraMetadata.CONTROL_AF_TRIGGER_START);
-
-            if (isMeteringAreaAFSupported) {
-                int focusSize = 50;
-                MeteringRectangle focusAreaTouch = new MeteringRectangle(textureView.getWidth() / 2 - focusSize / 2,
-                        textureView.getHeight() / 2 - focusSize / 2,
-                        focusSize,
-                        focusSize,
-                        MeteringRectangle.METERING_WEIGHT_MAX - 1);
-                previewRequestBuilder.set(CaptureRequest.CONTROL_AF_REGIONS, new MeteringRectangle[]{focusAreaTouch});
-            }
-
-            // Tell #captureCallback to wait for the lock.
-            state = STATE_WAITING_LOCK;
-            previewRequestBuilder.setTag("LOCK_FOCUS");
-            captureSession.capture(previewRequestBuilder.build(), captureCallback, backgroundHandler);
-
-            previewRequestBuilder.set(CaptureRequest.CONTROL_AF_TRIGGER, CameraMetadata.CONTROL_AF_TRIGGER_IDLE);
-
-            previewRequestBuilder.setTag("IDLE");
-            captureSession.setRepeatingRequest(previewRequestBuilder.build(), captureCallback, backgroundHandler);
-        } catch (CameraAccessException e) {
-            e.printStackTrace();
-        }
-    }
-
-    /**
-     * Run the precapture sequence for capturing a still image. This method should be called when
-     * we get a response in {@link #captureCallback} from {@link #lockFocus()}.
-     */
-    private void runPrecaptureSequence() {
-        if (captureSession == null) {
-            return;
-        }
-        Log.d(TAG, "Run precapture sequence");
-        try {
-            // This is how to tell the camera to trigger.
-            previewRequestBuilder.set(CaptureRequest.CONTROL_AE_PRECAPTURE_TRIGGER,
-                    CaptureRequest.CONTROL_AE_PRECAPTURE_TRIGGER_START);
-            // Tell #mCaptureCallback to wait for the precapture sequence to be set.
-            state = STATE_WAITING_PRECAPTURE;
-            previewRequestBuilder.setTag("PRE_CAPTURE");
-            captureSession.capture(previewRequestBuilder.build(), captureCallback,
-                    backgroundHandler);
-        } catch (CameraAccessException e) {
-            e.printStackTrace();
-        }
-    }
-
-    /**
-     * Capture a still picture. This method should be called when we get a response in
-     * {@link #captureCallback} from both {@link #lockFocus()}.
-     */
-    private void captureStillPicture() {
-        Log.d(TAG, "captureStillPicture");
-        try {
-            if (null == activity || null == captureSession || null == cameraDevice) {
-                return;
-            }
-            // This is the CaptureRequest.Builder that we use to take a picture.
-            final CaptureRequest.Builder captureBuilder =
-                    cameraDevice.createCaptureRequest(CameraDevice.TEMPLATE_STILL_CAPTURE);
-
-            captureBuilder.addTarget(stillReader.getSurface());
-            captureBuilder.addTarget(new Surface(textureView.getSurfaceTexture()));
-
-            // Use the same AE and AF modes as the preview.
-            captureBuilder.set(CaptureRequest.CONTROL_AF_MODE,
-                    CaptureRequest.CONTROL_AF_MODE_CONTINUOUS_PICTURE);
-            setFlashOnRequestBuilder(captureBuilder);
-
-            CameraCaptureSession.CaptureCallback captureCallback
-                    = new CameraCaptureSession.CaptureCallback() {
-
-                @Override
-                public void onCaptureCompleted(CameraCaptureSession session,
-                                               CaptureRequest request,
-                                               TotalCaptureResult result) {
-                    Log.d(TAG, "Still camera captureCallback");
-                    unlockFocus();
-                }
-            };
-            captureSession.capture(captureBuilder.build(), captureCallback, backgroundHandler);
-        } catch (CameraAccessException e) {
-            e.printStackTrace();
-        }
-    }
-
-    /**
-     * Unlock the focus. This method should be called when still image capture sequence is
-     * finished.
-     */
-    private void unlockFocus() {
-        if (captureSession == null) {
-            return;
-        }
-        Log.d(TAG, "unlock focus");
-        try {
-            // Reset the auto-focus trigger
-            previewRequestBuilder.set(CaptureRequest.CONTROL_AF_TRIGGER,
-                    CameraMetadata.CONTROL_AF_TRIGGER_CANCEL);
-            previewRequestBuilder.setTag("UNLOCK_FOCUS");
-            captureSession.capture(previewRequestBuilder.build(), captureCallback,
-                    backgroundHandler);
-
-            // After this, the camera will go back to the normal state of preview.
-            state = STATE_PREVIEW;
-
-            captureSession.setRepeatingRequest(previewRequest, captureCallback,
-                    backgroundHandler);
-        } catch (CameraAccessException e) {
-            e.printStackTrace();
-        }
-    }
-
-    /**
-     * Configures the necessary {@link Matrix} transformation to `mTextureView`. This method should be
-     * called after the camera preview size is determined in setUpCameraOutputs and also the size of
-     * `mTextureView` is fixed.
-     *
-     * @param viewWidth The width of `mTextureView`
-     * @param viewHeight The height of `mTextureView`
-     */
-    private void configureTransform(final int viewWidth, final int viewHeight) {
-        if (null == textureView || null == previewSize || null == activity) {
-            return;
-        }
-        final int rotation = activity.getWindowManager().getDefaultDisplay().getRotation();
-        final Matrix matrix = new Matrix();
-        final RectF viewRect = new RectF(0, 0, viewWidth, viewHeight);
-        final RectF bufferRect = new RectF(0, 0, previewSize.getHeight(), previewSize.getWidth());
-        final float centerX = viewRect.centerX();
-        final float centerY = viewRect.centerY();
-        if (Surface.ROTATION_90 == rotation || Surface.ROTATION_270 == rotation) {
-            bufferRect.offset(centerX - bufferRect.centerX(), centerY - bufferRect.centerY());
-            matrix.setRectToRect(viewRect, bufferRect, Matrix.ScaleToFit.FILL);
-            final float scale =
-                    Math.max(
-                            (float) viewHeight / previewSize.getHeight(),
-                            (float) viewWidth / previewSize.getWidth());
-            matrix.postScale(scale, scale, centerX, centerY);
-            matrix.postRotate(90 * (rotation - 2), centerX, centerY);
-        } else if (Surface.ROTATION_180 == rotation) {
-            matrix.postRotate(180, centerX, centerY);
-        }
-        textureView.setTransform(matrix);
-    }
-
-    /**
-     * Callback for Activities to use to initialize their data once the selected preview size is
-     * known.
-     */
-    public interface ConnectionCallback {
-        void onPreviewSizeChosen(Size previewSize, Size stillSize, int cameraRotation, boolean supportsTorchMode);
-    }
 
     /** Compares two {@code Size}s based on their areas. */
     static class CompareSizesByArea implements Comparator<Size> {
@@ -789,33 +233,6 @@ public class CameraController {
             // We cast here to ensure the multiplications won't overflow
             return Long.signum(
                     (long) lhs.getWidth() * lhs.getHeight() - (long) rhs.getWidth() * rhs.getHeight());
-        }
-    }
-
-    private void setFlashOnRequestBuilder(CaptureRequest.Builder requestBuilder) {
-        if (requestBuilder != null) {
-            if (flashEnabled) {
-                requestBuilder.set(CaptureRequest.FLASH_MODE, CaptureRequest.FLASH_MODE_TORCH);
-
-            } else {
-                requestBuilder.set(CaptureRequest.FLASH_MODE, CaptureRequest.FLASH_MODE_OFF);
-            }
-        }
-    }
-
-    public void setFlashEnabled(boolean flashEnabled) {
-        if (captureSession == null) {
-            return;
-        }
-        this.flashEnabled = flashEnabled;
-        if (previewRequestBuilder != null) {
-            setFlashOnRequestBuilder(previewRequestBuilder);
-            previewRequest = previewRequestBuilder.build();
-            try {
-                captureSession.setRepeatingRequest(previewRequest, captureCallback, backgroundHandler);
-            } catch (CameraAccessException e) {
-                e.printStackTrace();
-            }
         }
     }
 }
