+++ FluStudy_us/android/app/src/main/java/host/exp/exponent/DetectorView.java
@@ -10,13 +10,11 @@ import android.content.Context;
 import android.content.pm.PackageManager;
 import android.graphics.Bitmap;
 import android.graphics.Bitmap.Config;
+import android.graphics.BitmapFactory;
 import android.graphics.Canvas;
 import android.graphics.Matrix;
 import android.graphics.RectF;
-import android.hardware.camera2.CameraAccessException;
-import android.hardware.camera2.CameraCharacteristics;
-import android.hardware.camera2.CameraManager;
-import android.hardware.camera2.params.StreamConfigurationMap;
+import android.hardware.Camera;
 import android.media.Image;
 import android.media.ImageReader;
 import android.net.Uri;
@@ -119,7 +117,8 @@ public class DetectorView extends LinearLayout implements
         this.detectorListener = listener;
     }
 
-    public void onPreviewSizeChosen(final Size previewSize, final Size stillSize, final int rotation, boolean supportsTorchMode) {
+    public void onPreviewSizeChosen(final Size previewSize, final Size stillSize,
+                                    final int rotation, boolean supportsTorchMode) {
         this.previewSize = previewSize;
         this.stillSize = stillSize;
 
@@ -128,10 +127,11 @@ public class DetectorView extends LinearLayout implements
         screenWidth = getWidth();
 
         Log.i(TAG, "Camera orientation relative to screen canvas: " + sensorOrientation);
-        Log.i(TAG, "Initializing at preview size " + previewSize.getWidth() + ", " + previewSize.getHeight());
-        Log.i(TAG, "Initializing at still size " + stillSize.getWidth() + ", " + stillSize.getHeight());
+        Log.i(TAG, "Initializing at preview size " + previewSize);
+        Log.i(TAG, "Initializing at still size " + stillSize);
 
-        detectorListener.onRDTCameraReady(supportsTorchMode, screenWidth, screenHeight);
+        detectorListener.onRDTCameraReady(supportsTorchMode, screenWidth, screenHeight,
+                cameraController instanceof LegacyRDTCamera);
     }
 
     protected int getScreenOrientation() {
@@ -211,9 +211,12 @@ public class DetectorView extends LinearLayout implements
             vBuffer.get(vBytes);
         }
 
-
-        private void updateBitmaps() {
-            imageBitmap.setPixels(getRgbBytes(), 0, imageWidth, 0, 0, imageWidth, imageHeight);
+        private void updateBitmaps(Bitmap candidateImageBitmap) {
+            if (candidateImageBitmap == null) {
+                imageBitmap.setPixels(getRgbBytes(), 0, imageWidth, 0, 0, imageWidth, imageHeight);
+            } else {
+                imageBitmap = candidateImageBitmap;
+            }
             final Canvas canvas = new Canvas(boxModelBitmap);
             canvas.drawBitmap(imageBitmap, imageToModelTransform, null);
             readyForNextImage();
@@ -238,6 +241,7 @@ public class DetectorView extends LinearLayout implements
             initialized = true;
         }
 
+        /** Callback for Camera2 API */
         @Override
         public void onImageAvailable(ImageReader reader) {
             // We need to wait until we have some size from onPreviewSizeChosen
@@ -296,12 +300,60 @@ public class DetectorView extends LinearLayout implements
                                 isProcessingFrame = false;
                             }
                         };
-                processImage();
+                processImage(null);
             } catch (final Exception e) {
                 Log.e(TAG, "Exception in preview onImageAvailable: " + e.toString());
             } finally {
                 Trace.endSection(); // ImageAvailable
             }
+        }
+
+        /** Callback for android.hardware.Camera API */
+        protected void onFrame(byte[] bytes, Camera camera, boolean isStill) {
+            if (previewSize == null) {
+                Camera.Size currentPreviewSize = camera.getParameters().getPreviewSize();
+                Camera.Size stillFrameSize = camera.getParameters().getPictureSize();
+                onPreviewSizeChosen(
+                        new Size(currentPreviewSize.width, currentPreviewSize.height),
+                        // Legacy camera will use same size for preview as still image
+                        new Size(stillFrameSize.width, stillFrameSize.height),
+                        90, false);
+            }
+
+            if (!initialized) {
+                initialize();
+            }
+
+            if (!resourceLoader.openCVReady()) {
+                return;
+            }
+
+            if (isProcessingFrame) {
+                return;
+            }
+
+            isProcessingFrame = true;
+
+            if (!isStill) {
+                imageConverter =
+                        new Runnable() {
+                            @Override
+                            public void run() {
+                                ImageUtils.convertYUV420SPToARGB8888(
+                                        bytes, imageWidth, imageHeight, rgbBytes);
+                            }
+                        };
+            }
+
+            postInferenceCallback =
+                    new Runnable() {
+                        @Override
+                        public void run() {
+                            camera.addCallbackBuffer(bytes);
+                            isProcessingFrame = false;
+                        }
+                    };
+            processImage(isStill ? BitmapFactory.decodeByteArray(bytes, 0, bytes.length) :  null);
         }
 
         protected List<Classifier.Recognition> runPhaseOne() {
@@ -314,7 +366,7 @@ public class DetectorView extends LinearLayout implements
             return filterResults(BOX_MINIMUM_CONFIDENCE_TF_OD_API, results, true);
         }
 
-        private void processImage() {
+        private void processImage(Bitmap candidateImageBitmap) {
             // No mutex needed as this method is not reentrant.
             if (analyzingFrame) {
                 readyForNextImage();
@@ -324,7 +376,7 @@ public class DetectorView extends LinearLayout implements
 
             Trace.beginSection("processImage");
 
-            updateBitmaps();
+            updateBitmaps(candidateImageBitmap);
 
             runInBackground(
                     new Runnable() {
@@ -398,7 +450,7 @@ public class DetectorView extends LinearLayout implements
         }
     }
 
-    private class PreviewImageListener extends ImageListener {
+    public class PreviewImageListener extends ImageListener implements Camera.PreviewCallback {
 
         protected void initialize() {
             imageWidth = previewSize.getWidth();
@@ -456,9 +508,16 @@ public class DetectorView extends LinearLayout implements
 
             detectorListener.onRDTDetected(iprdResult, rdtResult, failureReason);
         }
+
+
+        /** Callback for android.hardware.Camera API */
+        @Override
+        public void onPreviewFrame(byte[] bytes, Camera camera) {
+            super.onFrame(bytes, camera, false);
+        }
     }
 
-    private class StillImageListener extends ImageListener {
+    public class StillImageListener extends ImageListener implements Camera.PictureCallback {
         protected void initialize() {
             imageWidth = stillSize.getWidth();
             imageHeight = stillSize.getHeight();
@@ -498,34 +557,11 @@ public class DetectorView extends LinearLayout implements
             }
             stillCaptureInProgress = false;
         }
-    }
 
-    private String chooseCamera() {
-        final CameraManager manager = (CameraManager) activity.getSystemService(Context.CAMERA_SERVICE);
-        try {
-            for (final String cameraId : manager.getCameraIdList()) {
-                final CameraCharacteristics characteristics = manager.getCameraCharacteristics(cameraId);
-
-                // We don't use a front facing camera in this sample.
-                final Integer facing = characteristics.get(CameraCharacteristics.LENS_FACING);
-                if (facing != null && facing == CameraCharacteristics.LENS_FACING_FRONT) {
-                    continue;
-                }
-
-                final StreamConfigurationMap map =
-                        characteristics.get(CameraCharacteristics.SCALER_STREAM_CONFIGURATION_MAP);
-
-                if (map == null) {
-                    continue;
-                }
-
-                return cameraId;
-            }
-        } catch (CameraAccessException e) {
-            Log.e(TAG, "Not allowed to access camera");
+        @Override
+        public void onPictureTaken(byte[] bytes, Camera camera) {
+            onFrame(bytes, camera, true);
         }
-
-        return null;
     }
 
     private boolean hasPermission() {
@@ -551,8 +587,7 @@ public class DetectorView extends LinearLayout implements
     }
 
     protected void initCameraController() {
-        String cameraId = chooseCamera();
-        cameraController = new CameraController(
+        cameraController = CameraController.getCamera(
                 activity,
                 textureView,
                 this,
@@ -560,7 +595,6 @@ public class DetectorView extends LinearLayout implements
                 new StillImageListener(),
                 getDesiredPreviewFrameSize());
 
-        cameraController.setCamera(cameraId);
     }
 
     public synchronized void onResume() {
@@ -627,7 +661,7 @@ public class DetectorView extends LinearLayout implements
     }
 
     public interface DetectorListener {
-        void onRDTCameraReady(boolean supportsTorchMode, int screenWidth, int screenHeight);
+        void onRDTCameraReady(boolean supportsTorchMode, int screenWidth, int screenHeight, boolean legacyCamera);
         void onRDTDetected(
                 IprdAdapter.Result iprdResult,
                 RDTTracker.RDTResult rdtResult,
