+++ FluStudy_us/android/app/src/main/java/host/exp/exponent/LegacyRDTCamera.java
@@ -1,0 +1,166 @@
+package host.exp.exponent;
+
+import android.app.Activity;
+import android.graphics.SurfaceTexture;
+import android.hardware.Camera;
+import android.hardware.Camera.CameraInfo;
+import android.os.Handler;
+import android.os.HandlerThread;
+import android.util.Log;
+import android.util.Size;
+import android.view.TextureView;
+
+import java.io.IOException;
+import java.util.List;
+
+import host.exp.exponent.customview.AutoFitTextureView;
+import host.exp.exponent.env.ImageUtils;
+
+public class LegacyRDTCamera extends CameraController {
+    private static final String TAG = "LegacyCamera";
+
+    private Camera camera;
+
+    /**
+     * {@link TextureView.SurfaceTextureListener} handles several lifecycle events on a {@link
+     * TextureView}.
+     */
+    private final TextureView.SurfaceTextureListener surfaceTextureListener =
+            new TextureView.SurfaceTextureListener() {
+                @Override
+                public void onSurfaceTextureAvailable(
+                        final SurfaceTexture texture, final int width, final int height) {
+
+                    int index = getCameraId();
+                    camera = Camera.open(index);
+
+                    try {
+                        Camera.Parameters parameters = camera.getParameters();
+                        List<String> focusModes = parameters.getSupportedFocusModes();
+                        if (focusModes != null
+                                && focusModes.contains(Camera.Parameters.FOCUS_MODE_CONTINUOUS_PICTURE)) {
+                            parameters.setFocusMode(Camera.Parameters.FOCUS_MODE_CONTINUOUS_PICTURE);
+                        }
+                        List<Camera.Size> cameraSizes = parameters.getSupportedPreviewSizes();
+                        Size[] sizes = new Size[cameraSizes.size()];
+                        int i = 0;
+                        for (Camera.Size size : cameraSizes) {
+                            sizes[i++] = new Size(size.width, size.height);
+                        }
+                        Size previewSize =
+                                chooseOptimalSize(sizes, desiredSize.getWidth(), desiredSize.getHeight());
+                        parameters.setPreviewSize(previewSize.getWidth(), previewSize.getHeight());
+                        camera.setDisplayOrientation(90);
+                        camera.setParameters(parameters);
+                        camera.setPreviewTexture(texture);
+                    } catch (IOException exception) {
+                        camera.release();
+                    }
+
+                    camera.setPreviewCallbackWithBuffer(imageListener);
+                    Camera.Size s = camera.getParameters().getPreviewSize();
+                    camera.addCallbackBuffer(new byte[ImageUtils.getYUVByteSize(s.height, s.width)]);
+
+                    textureView.setAspectRatio(s.height, s.width);
+
+                    camera.startPreview();
+                }
+
+                @Override
+                public void onSurfaceTextureSizeChanged(
+                        final SurfaceTexture texture, final int width, final int height) {}
+
+                @Override
+                public boolean onSurfaceTextureDestroyed(final SurfaceTexture texture) {
+                    return true;
+                }
+
+                @Override
+                public void onSurfaceTextureUpdated(final SurfaceTexture texture) {}
+            };
+
+    /** An additional thread for running tasks that shouldn't block the UI. */
+    private HandlerThread backgroundThread;
+
+    protected LegacyRDTCamera(final Activity activity,
+                     final AutoFitTextureView textureView,
+                     final ConnectionCallback callback,
+                     final DetectorView.PreviewImageListener imageListener,
+                     final DetectorView.StillImageListener stillImageListener,
+                     final Size desiredSize,
+                     final String cameraId) {
+        super(activity, textureView, callback, imageListener, stillImageListener, desiredSize,
+                cameraId);
+    }
+
+    @Override
+    public void onResume() {
+        startBackgroundThread();
+        // When the screen is turned off and turned back on, the SurfaceTexture is already
+        // available, and "onSurfaceTextureAvailable" will not be called. In that case, we can open
+        // a camera and start preview from here (otherwise, we wait until the surface is ready in
+        // the SurfaceTextureListener).
+
+        if (textureView.isAvailable()) {
+            camera.startPreview();
+        } else {
+            textureView.setSurfaceTextureListener(surfaceTextureListener);
+        }
+    }
+
+    @Override
+    public void onPause() {
+        stopCamera();
+        stopBackgroundThread();
+    }
+
+    @Override
+    public void captureStill() {
+        if (camera != null) {
+            camera.takePicture(null, null, null, stillImageListener);
+        }
+    }
+
+    @Override
+    public void setFlashEnabled(boolean flashEnabled) {
+        // We're just going to skip flash for legacy devices
+    }
+
+    /** Starts a background thread and its {@link Handler}. */
+    private void startBackgroundThread() {
+        backgroundThread = new HandlerThread("CameraBackground");
+        backgroundThread.start();
+    }
+
+    /** Stops the background thread and its {@link Handler}. */
+    private void stopBackgroundThread() {
+        if (backgroundThread == null) {
+            return;
+        }
+        backgroundThread.quitSafely();
+        try {
+            backgroundThread.join();
+            backgroundThread = null;
+        } catch (final InterruptedException e) {
+            Log.e(TAG, "Exception!\n" + e.getMessage());
+        }
+    }
+
+    protected void stopCamera() {
+        if (camera != null) {
+            camera.stopPreview();
+            camera.setPreviewCallback(null);
+            camera.release();
+            camera = null;
+        }
+    }
+
+    private int getCameraId() {
+        CameraInfo ci = new CameraInfo();
+        for (int i = 0; i < Camera.getNumberOfCameras(); i++) {
+            Camera.getCameraInfo(i, ci);
+            if (ci.facing == CameraInfo.CAMERA_FACING_BACK) return i;
+        }
+        return -1; // No camera found
+    }
+}
