+++ lib/chillsProtocol.ts
@@ -182,6 +182,7 @@ export interface RDTInfo {
   captureTime?: number;
   flashEnabled?: boolean;
   rdtTotalTime?: number;
+  legacyCameraApi?: boolean;
 
   // Other RDT reader information can go here:
   //  e.g. time taken to capture, length of time test strip was in solution, etc.
